push before battery dies
